requirements.txt
pylint
PyOpenGL
PyOpenGL-accelerate
PyOpenGL-Demo
PyQt6
numpy
